:truck: move capture lambda code
to a src directory and rename.  There will code for a second lambda (the trigger function) so this helps disambiguate
