+++ src/page-capture/validatePlugin.js
